+++ src/accelerate/commands/config/cluster.py
@@ -14,9 +14,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from accelerate.state import ComputeEnvironment, DistributedType
-
 from ...utils import is_deepspeed_available
+from ...utils.dataclasses import ComputeEnvironment, DistributedType
 from .config_args import ClusterConfig
 from .config_utils import _ask_field, _convert_distributed_mode, _convert_yes_no_to_bool
